Move validate_crate_name to rustc_metadata
