+++ src/librustc_codegen_utils/lib.rs
@@ -37,6 +37,7 @@ extern crate rustc_incremental;
 extern crate syntax;
 extern crate syntax_pos;
 #[macro_use] extern crate rustc_data_structures;
+extern crate rustc_metadata;
 
 use rustc::ty::TyCtxt;
